refactor: centralize system paths and SSH key discovery in config_observer
